rename fe project
