+++ frontend/app/components/home/Bar.js
