+++ frontend/app/components/home/SwitchClients.js
@@ -1,6 +1,6 @@
 import Link from "next/link";
 import React from "react";
-import { useCustomTheme } from "./GetTheme";
+import { useCustomTheme } from "../../hooks/useCustomTheme";
 
 const SwitchClients = () => {
   const resolvedTheme = useCustomTheme();
