+++ frontend/app/components/shared/Footer.js
