--- nextjs-clone/app/data/ClientResource.js
@@ -1,206 +1,0 @@
-export const ConsensusClientsData = [
-  {
-    name: "* Grandine",
-    clientLink: "https://sifrai.com/",
-    githubLink: "https://github.com/grandinetech/grandine",
-    docsLink: "https://github.com/grandinetech/grandine",
-    chatLink: "https://discord.com/invite/H9XCdUSyZd",
-    chatType: "discord",
-    status: "beta",
-    support: "Linux, Win, macOS",
-    language: "-",
-    donateLink: "",
-    donation: 1,
-  },
-  {
-    name: "Lighthouse",
-    clientLink: "https://lighthouse.sigmaprime.io/",
-    githubLink: "https://github.com/sigp/lighthouse/",
-    docsLink: "https://lighthouse-book.sigmaprime.io/",
-    chatLink: "https://discord.com/invite/cyAszAh",
-    chatType: "discord",
-    status: "stable",
-    support: "Linux, Win, macOS, ARM",
-    language: "Rust",
-    donateLink: "https://protocol-guild.readthedocs.io/en/latest/index.html",
-    donation: 1,
-  },
-  {
-    name: "Lodestar",
-    clientLink: "https://lodestar.chainsafe.io/",
-    githubLink: "https://github.com/ChainSafe/lodestar",
-    docsLink: "https://hackmd.io/@philknows/rk5cDvKmK",
-    chatLink: "https://discord.com/invite/yjyvFRP",
-    chatType: "discord",
-    status: "stable",
-    support: "Linux, Win, macOS",
-    language: "TypeScript",
-    donateLink: "https://protocol-guild.readthedocs.io/en/latest/index.html",
-    donation: 1,
-  },
-  {
-    name: "Nimbus",
-    clientLink: "https://nimbus.team/",
-    githubLink: "https://github.com/status-im/nimbus-eth2",
-    docsLink: "https://nimbus.team/docs/",
-    chatLink: "https://discord.com/invite/qnjVyhatUa",
-    chatType: "discord",
-    status: "stable",
-    support: "Linux, Win, macOS, ARM",
-    language: "Nim",
-    donateLink: "",
-    donation: 1,
-  },
-  {
-    name: "Prysm",
-    clientLink: "https://prysmaticlabs.com/",
-    githubLink: "https://github.com/prysmaticlabs/prysm",
-    docsLink: "https://docs.prylabs.network/docs/getting-started",
-    chatLink: "https://discord.com/invite/YMVYzv6",
-    chatType: "discord",
-    status: "stable",
-    support: "Linux, Win, macOS, ARM",
-    language: "Golang",
-    donateLink: "https://protocol-guild.readthedocs.io/en/latest/index.html",
-    donation: 1,
-  },
-  {
-    name: "Teku",
-    clientLink: "https://consensys.io/teku",
-    githubLink: "https://github.com/ConsenSys/teku",
-    docsLink: "https://docs.teku.consensys.io/",
-    chatLink: "https://discord.com/invite/9mCVSY6",
-    chatType: "discord",
-    status: "stable",
-    support: "Linux, Win, macOS",
-    language: "Java",
-    donateLink: "https://protocol-guild.readthedocs.io/en/latest/index.html",
-    donation: 1,
-  },
-];
-
-export const ExecutionClientsData = [
-  {
-    name: "Akula",
-    clientLink: "",
-    githubLink: "https://github.com/akula-bft/akula",
-    docsLink: "https://github.com/akula-bft/akula#running",
-    chatLink: "https://t.me/akula_bft",
-    chatType: "telegram",
-    status: "deprecated",
-    support: "-",
-    language: "-",
-    donateLink:
-      "https://etherscan.io/address/0x048a9ea34FcE0B2828a2703e486648c21ACCCC1A",
-    donation: 2,
-  },
-  {
-    name: "Besu",
-    clientLink: "https://www.hyperledger.org/projects/besu",
-    githubLink: "https://github.com/hyperledger/besu/",
-    docsLink: "https://besu.hyperledger.org/",
-    chatLink: "https://discord.com/invite/hyperledger",
-    chatType: "discord",
-    status: "stable",
-    support: "Linux, Win, macOS",
-    language: "Java",
-    donateLink: "https://protocol-guild.readthedocs.io/en/latest/index.html",
-    donation: 1,
-  },
-  {
-    name: "Erigon",
-    clientLink: "https://github.com/ledgerwatch/erigon#erigon",
-    githubLink: "https://github.com/ledgerwatch/erigon",
-    docsLink: "https://github.com/ledgerwatch/erigon#usage",
-    chatLink: "",
-    chatType: "",
-    status: "alpha & beta",
-    support: "Linux, Win, macOS, ARM",
-    language: "Golang",
-    donateLink:
-      "https://etherscan.io/address/0x8BFBB529A9E85fDC4b70A4FCdC0D68Bb298B8816",
-    donation: 2,
-  },
-  {
-    name: "EthereumJS",
-    clientLink: "https://ethereumjs.github.io/",
-    githubLink:
-      "https://github.com/ethereumjs/ethereumjs-monorepo/tree/master/packages/client",
-    docsLink:
-      "https://github.com/ethereumjs/ethereumjs-monorepo/tree/master/packages/client/docs",
-    chatLink: "https://discord.com/invite/TNwARpR",
-    chatType: "discord",
-    status: "alpha",
-    support: "Linux, Win, macOS",
-    language: "TypeScript",
-    donateLink: "",
-    donation: 0,
-  },
-  {
-    name: "Geth",
-    clientLink: "https://geth.ethereum.org/",
-    githubLink: "https://github.com/ethereum/go-ethereum",
-    docsLink: "https://geth.ethereum.org/docs",
-    chatLink: "https://discord.com/invite/nthXNEv",
-    chatType: "discord",
-    status: "stable",
-    support: "Linux, Win, macOS, ARM",
-    language: "Golang",
-    donateLink: "https://protocol-guild.readthedocs.io/en/latest/index.html",
-    donation: 1,
-  },
-  {
-    name: "Nethermind",
-    clientLink: "https://www.nethermind.io/",
-    githubLink: "https://github.com/NethermindEth/nethermind",
-    docsLink:
-      "https://docs.nethermind.io/nethermind/first-steps-with-nethermind/getting-started",
-    chatLink: "https://discord.com/invite/PaCMRFdvWT",
-    chatType: "discord",
-    status: "stable",
-    support: "Linux, Win, macOS, ARM",
-    language: ".NET",
-    donateLink: "https://protocol-guild.readthedocs.io/en/latest/index.html",
-    donation: 1,
-  },
-  {
-    name: "Nimbus",
-    clientLink: "https://nimbus.team/",
-    githubLink: "https://github.com/status-im/nimbus-eth1",
-    docsLink: "https://github.com/status-im/nimbus-eth1/tree/master/doc",
-    chatLink: "https://discord.com/invite/qnjVyhatUa",
-    chatType: "discord",
-    status: "pre-alpha",
-    support: "-",
-    language: "Nim",
-    donateLink: "",
-    donation: 0,
-  },
-  {
-    name: "Reth",
-    clientLink: "https://paradigmxyz.github.io/reth/",
-    githubLink: "https://github.com/paradigmxyz/reth",
-    docsLink: "https://paradigmxyz.github.io/reth/",
-    chatLink: "https://t.me/paradigm_reth",
-    chatType: "telegram",
-    status: "alpha",
-    support: "Linux, Win, macOS, ARM",
-    language: "Rust",
-    donateLink: "",
-    donation: 0,
-  },
-  {
-    name: "Silkworm",
-    clientLink: "https://github.com/erigontech/silkworm#about-silkworm",
-    githubLink: "https://github.com/erigontech/silkworm",
-    docsLink: "https://github.com/erigontech/silkworm#source",
-    chatLink: "",
-    chatType: "",
-    status: "pre-alpha",
-    support: "pre-alpha",
-    language: "C++",
-    donateLink:
-      "https://etherscan.io/address/0x8BFBB529A9E85fDC4b70A4FCdC0D68Bb298B8816",
-    donation: 2,
-  },
-];
